Revert "Add/BBE generic feedback section (#69072)"
This reverts commit 936b5f138cb34c37d643423cc1d17e92fd12cd84.
