+++ client/signup/steps/website-content/page-details.tsx
@@ -5,34 +5,30 @@ import {
 	Label,
 	TextAreaField,
 	HorizontalGrid,
-	ContactInformation,
 } from 'calypso/signup/accordion-form/form-components';
 import { ValidationErrors } from 'calypso/signup/accordion-form/types';
 import {
-	MediaUploadData,
-	WordpressMediaUpload,
-} from 'calypso/signup/steps/website-content/wordpress-media-upload';
-import {
-	websiteContentFieldChanged,
-	imageUploaded,
 	imageRemoved,
+	imageUploaded,
 	imageUploadFailed,
 	imageUploadInitiated,
+	websiteContentFieldChanged,
 } from 'calypso/state/signup/steps/website-content/actions';
 import { getSelectedSite } from 'calypso/state/ui/selectors';
-import type { ContactPageData } from 'calypso/state/signup/steps/website-content/schema';
+import { MediaUploadData, WordpressMediaUpload } from './wordpress-media-upload';
+import type { PageData } from 'calypso/state/signup/steps/website-content/schema';
 import type { TranslateResult } from 'i18n-calypso';
 
 export const CONTENT_SUFFIX = 'Content';
 export const IMAGE_PREFIX = 'Image';
 
-export function ContactPageDetails( {
+export function PageDetails( {
 	page,
 	formErrors,
 	label,
 	onChangeField,
 }: {
-	page: ContactPageData;
+	page: PageData;
 	formErrors: ValidationErrors;
 	label: TranslateResult | undefined;
 	onChangeField?: ( { target: { name, value } }: ChangeEvent< HTMLInputElement > ) => void;
@@ -84,14 +80,14 @@ export function ContactPageDetails( {
 		);
 	};
 
-	const onFieldChanged = ( e: ChangeEvent< HTMLInputElement > ) => {
+	const onContentChange = ( e: ChangeEvent< HTMLInputElement > ) => {
 		const {
-			target: { name, value },
+			target: { value },
 		} = e;
 		dispatch(
 			websiteContentFieldChanged( {
 				pageId: page.id,
-				fieldName: name,
+				fieldName: 'content',
 				fieldValue: value,
 			} )
 		);
@@ -101,8 +97,8 @@ export function ContactPageDetails( {
 	return (
 		<>
 			<TextAreaField
-				name="content"
-				onChange={ onFieldChanged }
+				name={ fieldName }
+				onChange={ onContentChange }
 				value={ page.content }
 				error={ formErrors[ fieldName ] }
 				label={
@@ -111,21 +107,6 @@ export function ContactPageDetails( {
 						args: { pageTitle },
 					} )
 				}
-			/>
-			<ContactInformation
-				displayEmailProps={ {
-					value: page.displayEmail || '',
-					name: 'displayEmail',
-				} }
-				displayPhoneProps={ {
-					value: page.displayPhone || '',
-					name: 'displayPhone',
-				} }
-				displayAddressProps={ {
-					value: page.displayAddress || '',
-					name: 'displayAddress',
-				} }
-				onChange={ onFieldChanged }
 			/>
 			<Label>
 				{ translate( 'Upload up to 3 images to be used on your %(pageTitle)s page.', {
