--- client/signup/steps/website-content/section-types/index.tsx
@@ -1,6 +1,0 @@
-import { ContactPageDetails } from 'calypso/signup/steps/website-content/section-types/contact-page-details';
-import { DefaultPageDetails } from 'calypso/signup/steps/website-content/section-types/default-page-details';
-import { FeedbackSection } from 'calypso/signup/steps/website-content/section-types/feedback-section';
-import { LogoUploadSection } from 'calypso/signup/steps/website-content/section-types/logo-upload-section';
-
-export { ContactPageDetails, LogoUploadSection, FeedbackSection, DefaultPageDetails };
