+++ client/state/action-types.ts
@@ -857,8 +857,6 @@ export const SIGNUP_STEPS_SOCIAL_PROFILES_UPDATE = 'SIGNUP_STEPS_SOCIAL_PROFILES
 export const SIGNUP_STEPS_SOCIAL_PROFILES_RESET = 'SIGNUP_STEPS_SOCIAL_PROFILES_RESET';
 export const SIGNUP_STEPS_WEBSITE_CONTENT_REMOVE_LOGO_URL =
 	'SIGNUP_STEPS_WEBSITE_CONTENT_REMOVE_LOGO_URL';
-export const SIGNUP_STEPS_WEBSITE_CONTENT_FEEDBACK_CHANGE =
-	'SIGNUP_STEPS_WEBSITE_CONTENT_FEEDBACK_CHANGE';
 
 export const SIGNUP_STEPS_USER_EXPERIENCE_SET = 'SIGNUP_STEPS_USER_EXPERIENCE_SET';
 export const SITE_ADDRESS_AVAILABILITY_ERROR = 'SITE_ADDRESS_AVAILABILITY_ERROR';
