+++ client/state/signup/steps/website-content/schema.ts
@@ -72,11 +72,7 @@ export interface ContactPageData extends PageData {
 	displayAddress?: string;
 }
 
-export type WebsiteContent = {
-	pages: Array< PageData >;
-	siteLogoSection: { siteLogoUrl: string };
-	feedbackSection: { genericFeedback: string };
-};
+export type WebsiteContent = { pages: Array< PageData >; siteLogoUrl: string };
 export interface WebsiteContentCollection {
 	currentIndex: number;
 	websiteContent: WebsiteContent;
@@ -85,10 +81,6 @@ export interface WebsiteContentCollection {
 
 export const initialState: WebsiteContentCollection = {
 	currentIndex: 0,
-	websiteContent: {
-		pages: [],
-		siteLogoSection: { siteLogoUrl: '' },
-		feedbackSection: { genericFeedback: '' },
-	},
+	websiteContent: { pages: [], siteLogoUrl: '' },
 	imageUploadStates: {},
 };
