+++ client/state/signup/steps/website-content/test/selectors.ts
@@ -10,8 +10,7 @@ describe( 'selectors', () => {
 	test( 'should return the initial state as a default state', () => {
 		expect( getWebsiteContent( { signup: undefined } ) ).toEqual( {
 			pages: [],
-			siteLogoSection: { siteLogoUrl: '' },
-			feedbackSection: { genericFeedback: '' },
+			siteLogoUrl: '',
 		} );
 	} );
